+++ src/Frame/Frame.cc
@@ -2390,17 +2390,20 @@ bool Frame::GetStokesTypeIndex(const string& coordinate, int& stokes_index) {
 
     auto stokes_type = Stokes::Get(stokes_string);
     if (stokes_type) {
-        if (_loader->GetStokesTypeIndex(stokes_type, stokes_index)) {
-            stokes_ok = true;
-        } else if (Stokes::IsComputed(stokes_type)) {
+        if (Stokes::IsComputed(stokes_type)) {
             stokes_index = stokes_type;
             stokes_ok = true;
         } else {
-            int assumed_stokes_index = (stokes_type - 1) % 4;
-            if (NumStokes() > assumed_stokes_index) {
-                stokes_index = assumed_stokes_index;
-                stokes_ok = true;
-                spdlog::warn("Can not get stokes index from the header. Assuming stokes {} index is {}.", stokes_string, stokes_index);
+            if (!_loader->GetStokesIndices().empty()) {
+                // Stokes are defined in image
+                stokes_ok = _loader->GetStokesTypeIndex(stokes_type, stokes_index);
+            } else {
+                int assumed_stokes_index = (stokes_type - 1) % 4;
+                if (NumStokes() > assumed_stokes_index) {
+                    stokes_index = assumed_stokes_index;
+                    stokes_ok = true;
+                    spdlog::warn("Can not get stokes index from the header. Assuming stokes {} index is {}.", stokes_string, stokes_index);
+                }
             }
         }
     }
@@ -2510,28 +2513,21 @@ bool Frame::DoVectorFieldCalculation(const std::function<void(CARTA::VectorOverl
     int mip = _vector_field.Mip();
     bool fractional = _vector_field.Fractional();
     float threshold = _vector_field.Threshold();
-    bool calculate_pi = _vector_field.CalculatePi();
-    bool calculate_pa = _vector_field.CalculatePa();
-    bool current_stokes_as_pi = _vector_field.CurrStokesAsPi();
-    bool current_stokes_as_pa = _vector_field.CurrStokesAsPa();
-
-    // Get tiles
-    std::vector<Tile> tiles;
-    GetTiles(_dims.width, _dims.height, mip, tiles);
+    bool calculate_pi_pa = _vector_field.CalculatePi() || _vector_field.CalculatePa();
+    bool current_stokes_pi_pa = _vector_field.CurrStokesAsPi() || _vector_field.CurrStokesAsPa();
 
     // Initialize stokes maps for their flags, indices and data
     std::unordered_map<std::string, bool> stokes_flag{{"I", false}, {"Q", false}, {"U", false}};
     std::unordered_map<std::string, int> stokes_indices{{"I", -1}, {"Q", -1}, {"U", -1}};
 
     // Set stokes flags and get their indices
-    stokes_flag["I"] = (fractional || !std::isnan(threshold));
-    stokes_flag["Q"] = stokes_flag["U"] = (calculate_pi || calculate_pa);
-    for (auto one : stokes_flag) {
-        std::string stokes = one.first;
-        if (stokes_flag[stokes] && !GetStokesTypeIndex(stokes + "x", stokes_indices[stokes])) {
-            return false;
-        }
-    }
+    stokes_flag["I"] = (fractional || !std::isnan(threshold)) && GetStokesTypeIndex("I", stokes_indices["I"]);
+    stokes_flag["Q"] = calculate_pi_pa && GetStokesTypeIndex("Q", stokes_indices["Q"]);
+    stokes_flag["U"] = calculate_pi_pa && GetStokesTypeIndex("U", stokes_indices["U"]);
+
+    // Get tiles
+    std::vector<Tile> tiles;
+    GetTiles(_dims.width, _dims.height, mip, tiles);
 
     // Get image tiles data
     for (int i = 0; i < tiles.size(); ++i) {
@@ -2542,14 +2538,14 @@ bool Frame::DoVectorFieldCalculation(const std::function<void(CARTA::VectorOverl
         double progress = (double)(i + 1) / tiles.size();
 
         // Get current stokes data
-        if (current_stokes_as_pi || current_stokes_as_pa) {
+        if (current_stokes_pi_pa) {
             if (!GetDownsampledRasterData(stokes_data["CUR"], width, height, _z_index, CurrentStokes(), bounds, mip)) {
                 return false;
             }
         }
 
         // Get stokes data I, Q, or U
-        if (calculate_pi || calculate_pa) {
+        if (calculate_pi_pa) {
             for (auto one : stokes_flag) {
                 std::string stokes = one.first;
                 if (stokes_flag[stokes] &&